requirements.txt
torch>=2.0.0
torchvision
scikit-learn
Pillow
matplotlib
